Move existing target determinator to tools
This moves out the modulefinder determinator to `tools/testing` since it is supposed to be CI-only. This also simplifies run_test.py a little bit.

[ghstack-poisoned]
